Ajustado o componente do form de cadastro.
Transformado em componente e criado as requisições nos componentes pais.
