+++ client/src/components/selfRegister/index.tsx
@@ -5,7 +5,9 @@ import CadastroUsuario from '../../components/formRegisterUser';
 
 import Popup from "../popup";
 import axios from "axios";
+import { useNavigate } from 'react-router-dom';
 import { AiOutlineClose } from "react-icons/ai";
+import { FormData } from "../../utils/user";
 const apiUrl = process.env.REACT_APP_API_NODE_URL;
 
 interface ForgotPassPopupProps {
@@ -13,10 +15,41 @@ interface ForgotPassPopupProps {
 }
 
 const SelfRegisterPopup: React.FC<ForgotPassPopupProps> = ({ onClose }) => {
-  const [userEmail, setUserEmail] = useState("");
-  const [typeResetPass, setTypeResetPass] = useState("");
+  const [formData, setFormData] = useState<FormData>({
+    nome: '',
+    sobrenome: '',
+    username: '',
+    email: '',
+    password: '',
+    confpassword: ''
+  });
+  const navigate = useNavigate();
 
 
+  const handleSubmit = async (a: any) => {
+    //faz a requisiçao pro back
+    try {
+      const response = await axios.post(`${apiUrl}/createUser`, formData);
+
+      if (response.data.status === 'Ok') {
+        // Realiza o redirecionamento para outra página
+        navigate('/users');
+      }
+    }
+    catch (error) {
+      console.error(error);
+      alert("Ops... Algo deu errado ao efetuar o cadastro do usuário. Tente novamente");
+    }
+    // Limpa o formulário após o envio
+    setFormData({
+      nome: '',
+      sobrenome: '',
+      username: '',
+      email: '',
+      password: '',
+      confpassword: '',
+    });
+  };
 
   const onClosePopup = () => {
     onClose();
@@ -30,7 +63,7 @@ const SelfRegisterPopup: React.FC<ForgotPassPopupProps> = ({ onClose }) => {
             <AiOutlineClose />
           </button>
           <div>
-            <CadastroUsuario />
+            <CadastroUsuario formData={formData} setFormData={setFormData} onSubmit={handleSubmit} />
           </div>
 
         </div>
